requirements.txt
google-genai>=0.2.0
pillow>=9.0.0
tk; platform_system != "Windows" 